requirements.txt
asgiref==3.4.1
Django==3.2.16
Pillow==8.4.0
psycopg2==2.9.5
pytz==2022.6
sqlparse==0.4.3
typing_extensions==4.1.1
